requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
crispy-bootstrap5==2025.4
cssselect2==0.8.0
Django==4.2.20
django-crispy-forms==2.4
django-environ==0.12.0
django-formtools==2.5.1
fonttools==4.57.0
html5lib==1.1
pillow==11.1.0
pycparser==2.22
pydyf==0.6.0
pyphen==0.17.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.3.post1
